requirements.txt
anthropic
discord.py
python-dotenv
